Changed INDEX typedef to be int instead of long, which makes a lot more sense considering there is already a LONG typedef

+++ src/Scratch/Common.h
@@ -60,7 +60,7 @@
 #define TRUE 1
 #define FALSE 0
 
-typedef           long INDEX;
+typedef            int INDEX;
 typedef          float FLOAT;
 typedef         double DOUBLE;
 typedef            int BOOL;
